categories done!

+++ bookstore/shopping/delete-item.php
@@ -3,13 +3,29 @@
 
 <?php 
 
-if(isset($_POST['delete'])){
-    $id = $_POST['id'];
 
-    $udate = $conn->prepare("DELETE FROM cart WHERE id='$id'");
+    if ( $_SERVER['REQUEST_METHOD']=='GET' && realpath(__FILE__) == realpath( $_SERVER['SCRIPT_FILENAME'] ) ) {
+        /* 
+        Up to you which header to send, some prefer 404 even if 
+        the files does exist for security
+        */
+        header( 'HTTP/1.0 403 Forbidden', TRUE, 403 );
+        die(header('location: '.APPURL.'' ));
+    }
 
-    $udate->execute();
-}
+
+    if (!isset($_SESSION['username'])){
+        header("location: ".APPURL."");
+    }
+
+
+    if(isset($_POST['delete'])){
+        $id = $_POST['id'];
+
+        $udate = $conn->prepare("DELETE FROM cart WHERE id='$id'");
+
+        $udate->execute();
+    }
 
 ?>
